+++ packages/ui-mobile-playground/src/components/index.ts
@@ -1,2 +1,1 @@
 export * from './Playground';
-export * from './routes';